+++ backend/app/services/llm_services.py
@@ -2,7 +2,7 @@ from http.client import responses
 
 from openai import OpenAI
 
-from app.core.config import settings
+from backend.app.core.config import settings
 
 
 class LLMService: